+++ flask_express/__init__.py
@@ -1,105 +1,27 @@
 """
 created by: Aniket Sarkar(https://github.com/marktennyson)
-Please contribute in this project.
+Please contribute to this project.
 """
-from markupsafe import escape
-from markupsafe import Markup
-from werkzeug.exceptions import abort as abort
-from werkzeug.utils import redirect as redirect
-
-from flask import json as json
-from flask.app import Response as Response
-from flask.blueprints import Blueprint as Blueprint
-from flask.config import Config as Config
-from flask.ctx import after_this_request as after_this_request
-from flask.ctx import copy_current_request_context as copy_current_request_context
-from flask.ctx import has_app_context as has_app_context
-from flask.ctx import has_request_context as has_request_context
-from flask.globals import _app_ctx_stack as _app_ctx_stack
-from flask.globals import _request_ctx_stack as _request_ctx_stack
-from flask.globals import current_app as current_app
-from flask.globals import g as g
-from flask.globals import request as request
-from flask.globals import session as session
-from flask.helpers import flash as flash
-from flask.helpers import get_flashed_messages as get_flashed_messages
-from flask.helpers import get_template_attribute as get_template_attribute
-from flask.helpers import make_response as make_response
-from flask.helpers import safe_join as safe_join
-from flask.helpers import send_file as send_file
-from flask.helpers import send_from_directory as send_from_directory
-from flask.helpers import stream_with_context as stream_with_context
-from flask.helpers import url_for as url_for
-from flask.json import jsonify as jsonify
-from flask.signals import appcontext_popped as appcontext_popped
-from flask.signals import appcontext_pushed as appcontext_pushed
-from flask.signals import appcontext_tearing_down as appcontext_tearing_down
-from flask.signals import before_render_template as before_render_template
-from flask.signals import got_request_exception as got_request_exception
-from flask.signals import message_flashed as message_flashed
-from flask.signals import request_finished as request_finished
-from flask.signals import request_started as request_started
-from flask.signals import request_tearing_down as request_tearing_down
-from flask.signals import signals_available as signals_available
-from flask.signals import template_rendered as template_rendered
-from flask.templating import render_template as render_template
-from flask.templating import render_template_string as render_template_string
-
-
-from flask.scaffold import setupmethod
-from flask.scaffold import _endpoint_from_view_func                                
-from flask.app import Flask
-import typing as t
-from .request import Request
-from .response import Response as Responser
-from ._helper import get_main_ctx_view
 from os import path
 
+from werkzeug.datastructures import Headers
+from werkzeug.wrappers.response import Response as BaseResponse
 
-__all__ = (
-    "FlaskExpress",
-    "escape",
-    "Markup",
-    "abort",
-    "redirect",
-    "json",
-    "Response",
-    "Blueprint",
-    "Config",
-    "after_this_request",
-    "copy_current_request_context",
-    "has_app_context",
-    "has_request_context",
-    "_app_ctx_stack",
-    "_request_ctx_stack",
-    "current_app",
-    "g",
-    "request",
-    "session",
-    "flash",
-    "get_flashed_messages",
-    "get_template_attribute",
-    "make_response",
-    "safe_join",
-    "send_file",
-    "send_from_directory",
-    "stream_with_context",
-    "url_for",
-    "jsonify",
-    "appcontext_pushed",
-    "appcontext_popped",
-    "appcontext_tearing_down",
-    "before_render_template",
-    "got_request_exception",
-    "message_flashed",
-    "request_finished",
-    "request_tearing_down",
-    "signals_available",
-    "template_rendered",
-    "render_template",
-    "render_template_string",
-    "Responser",
-)
+from flask.scaffold import setupmethod
+from flask.json import jsonify
+from flask.globals import request as grequest
+from flask.scaffold import _endpoint_from_view_func                                
+from flask.app import Flask
+
+from .request import Request
+from .response import Response
+from ._helper import get_main_ctx_view
+
+import sys as sys
+import typing as t
+
+if t.TYPE_CHECKING:
+    from flask.typing import ResponseReturnValue
 
 
 class FlaskExpress(Flask):
@@ -129,6 +51,7 @@ class FlaskExpress(Flask):
                             instance_relative_config=instance_relative_config, 
                             static_host=static_host)
         self.config['ATTACHMENTS_FOLDER'] = path.join(path.abspath(path.dirname(self.import_name)), "attachments")
+
 
     @setupmethod
     def add_url_rule(
@@ -189,6 +112,116 @@ class FlaskExpress(Flask):
                 )
             view_func = get_main_ctx_view(view_func)
             self.view_functions[endpoint] = view_func
+
+
+    def make_response(self, rv: "ResponseReturnValue") -> Response:
+        """Convert the return value from a view function to an instance of
+        :attr:`response_class`.
+
+        :param rv: the return value from the view function. The view function
+            must return a response. Returning ``None``, or the view ending
+            without returning, is not allowed. The following types are allowed
+            for ``view_rv``:
+
+            ``str``
+                A response object is created with the string encoded to UTF-8
+                as the body.
+
+            ``bytes``
+                A response object is created with the bytes as the body.
+
+            ``dict``
+                A dictionary that will be jsonify'd before being returned.
+
+            ``tuple``
+                Either ``(body, status, headers)``, ``(body, status)``, or
+                ``(body, headers)``, where ``body`` is any of the other types
+                allowed here, ``status`` is a string or an integer, and
+                ``headers`` is a dictionary or a list of ``(key, value)``
+                tuples. If ``body`` is a :attr:`response_class` instance,
+                ``status`` overwrites the exiting value and ``headers`` are
+                extended.
+
+            :attr:`response_class`
+                The object is returned unchanged.
+
+            other :class:`~werkzeug.wrappers.Response` class
+                The object is coerced to :attr:`response_class`.
+
+            :func:`callable`
+                The function is called as a WSGI application. The result is
+                used to create a response object.
+
+        .. versionchanged:: 0.9
+           Previously a tuple was interpreted as the arguments for the
+           response object.
+        """
+        status = headers = None
+
+        # unpack tuple returns
+        if isinstance(rv, tuple):
+            len_rv = len(rv)
+
+            # a 3-tuple is unpacked directly
+            if len_rv == 3:
+                rv, status, headers = rv
+            # decide if a 2-tuple has status or headers
+            elif len_rv == 2:
+                if isinstance(rv[1], (Headers, dict, tuple, list)):
+                    rv, headers = rv
+                else:
+                    rv, status = rv
+            # other sized tuples are not allowed
+            else:
+                raise TypeError(
+                    "The view function did not return a valid response tuple."
+                    " The tuple must have the form (body, status, headers),"
+                    " (body, status), or (body, headers)."
+                )
+
+        # the body must not be None
+        if rv is None:
+            raise TypeError(
+                f"The view function for {grequest.endpoint!r} did not"
+                " return a valid response. The function either returned"
+                " None or ended without a return statement."
+            )
+
+        # make sure the body is an instance of the response class
+        if not isinstance(rv, self.response_class):
+            if isinstance(rv, (str, bytes, bytearray)):
+                # let the response class set the status and headers instead of
+                # waiting to do it manually, so that the class can handle any
+                # special logic
+                
+                rv = self.response_class().make_response(rv, status=status, headers=headers) # this requires a default type `Response` class.
+                status = headers = None
+            elif isinstance(rv, dict):
+                rv = jsonify(rv)
+            elif isinstance(rv, Response) or callable(rv):
+                # evaluate a WSGI callable, or coerce a different response
+                # class to the correct type
+                try:
+                    rv = self.response_class().force_type(rv, grequest.environ)  # type: ignore  # noqa: B950
+                except TypeError as e:
+                    raise TypeError(
+                        f"{e}\nThe view function did not return a valid"
+                        " response. The return type must be a string,"
+                        " dict, tuple, Response instance, or WSGI"
+                        f" callable, but it was a {type(rv).__name__}."
+                    ).with_traceback(sys.exc_info()[2])
+            else:
+                raise TypeError(
+                    "The view function did not return a valid"
+                    " response. The return type must be a string,"
+                    " dict, tuple, Response instance, or WSGI"
+                    f" callable, but it was a {type(rv).__name__}."
+                )
+
+        rv = t.cast(Response, rv)        
+        
+        return self.response_class().make_response_from_obj(rv)
+
 
     def listen(self, 
         port: t.Optional[int] = None,