+++ CHANGELOG.md
@@ -7,3 +7,7 @@
 - Some typo fixed at README.md.
 - Some typo fixed at Munch module.
 - Aded the proper logo.
+
+## 0.1.2
+- Fixed `flask-admin` related issue.
+- Added __MKDocs__ based documentation.